requirements.txt
aiohttp==3.9.2
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
certifi==2023.11.17
cfgv==3.4.0
discord.py==2.3.2
distlib==0.3.8
distro==1.9.0
filelock==3.13.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
identify==2.5.33
idna==3.6
multidict==6.0.4
nodeenv==1.8.0
openai==1.10.0
platformdirs==4.1.0
pre-commit==3.6.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.1
PyYAML==6.0.1
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.9.0
virtualenv==20.25.0
yarl==1.9.4
